Refactor numba_impl call
